+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to the "smart-log" extension will be documented in this file.
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
+## [1.0.0]
+
+- improved time regex / parsing
+- added example for "common log format" (http logs)
+- post time updates only on valid times
 
 ## [0.9.1]
 - Added telemetry using vscode-extension-telemetry with events: 'activate' and 'open file' (measurements number of fileConfigs).